[feature/#9] Create auction bid entity

+++ src/main/java/com/api/jaebichuri/auction/entity/AuctionBid.java
@@ -18,7 +18,7 @@ public class AuctionBid {
     private Long id;
 
     @Column(nullable = false)
-    private int bidPrice;
+    private Long bidPrice;
 
     @Column(nullable = false)
     private LocalDateTime bidTime;
